feat: move code in custom hook
